Type is a parser-level abstraction
