+++ rust/pattern/constraint/type_/plays.rs
@@ -24,7 +24,7 @@ use std::{fmt, iter};
 
 use crate::{
     common::{error::collect_err, token, validatable::Validatable, Result},
-    pattern::{variable::Reference, Type, TypeVariable, TypeVariableBuilder, UnboundVariable},
+    pattern::{variable::Reference, TypeVariable, TypeVariableBuilder, UnboundVariable},
     Label,
 };
 
@@ -36,7 +36,7 @@ pub struct PlaysConstraint {
 }
 
 impl PlaysConstraint {
-    fn new(role_type: TypeVariable, overridden_role_type: Option<TypeVariable>) -> Self {
+    pub(crate) fn new(role_type: TypeVariable, overridden_role_type: Option<TypeVariable>) -> Self {
         PlaysConstraint {
             relation_type: role_type
                 .label
@@ -112,18 +112,6 @@ impl From<(Label, Label)> for PlaysConstraint {
 impl From<UnboundVariable> for PlaysConstraint {
     fn from(role_type: UnboundVariable) -> Self {
         PlaysConstraint::from(role_type.into_type())
-    }
-}
-
-impl From<Type> for PlaysConstraint {
-    fn from(role_type: Type) -> Self {
-        PlaysConstraint::new(role_type.into_type_variable(), None)
-    }
-}
-
-impl From<(Type, Option<Type>)> for PlaysConstraint {
-    fn from((role_type, overridden_role_type): (Type, Option<Type>)) -> Self {
-        PlaysConstraint::new(role_type.into_type_variable(), overridden_role_type.map(Type::into_type_variable))
     }
 }
