+++ rust/pattern/constraint/type_/relates.rs
@@ -24,7 +24,7 @@ use std::{fmt, iter};
 
 use crate::{
     common::{error::collect_err, token, validatable::Validatable, Result},
-    pattern::{variable::Reference, Type, TypeVariable, TypeVariableBuilder, UnboundVariable},
+    pattern::{variable::Reference, TypeVariable, TypeVariableBuilder, UnboundVariable},
     Label,
 };
 
@@ -91,18 +91,6 @@ impl From<TypeVariable> for RelatesConstraint {
 impl From<(TypeVariable, Option<TypeVariable>)> for RelatesConstraint {
     fn from((role_type, overridden_role_type): (TypeVariable, Option<TypeVariable>)) -> Self {
         RelatesConstraint { role_type, overridden_role_type }
-    }
-}
-
-impl From<Type> for RelatesConstraint {
-    fn from(role_type: Type) -> Self {
-        RelatesConstraint::from(role_type.into_type_variable())
-    }
-}
-
-impl From<(Type, Option<Type>)> for RelatesConstraint {
-    fn from((role_type, overridden): (Type, Option<Type>)) -> Self {
-        RelatesConstraint::from((role_type.into_type_variable(), overridden.map(Type::into_type_variable)))
     }
 }
